Check access rule before updating class byte
checkCommand() shall be called before class byte is updated with the
channel number. Otherwise, valid trasmit request can be filtered out
becaue of the channel number.

Say that the APDU filter "00 55 00 00" is available and the current
channel number is 1, for example. The APDU header is changed to "01 xx
xx xx" if the class byte is updated by Util.setChannelToClassByte(). In
this case, the transmit request is incorrectly rejected because of the
channel number if we do not apply this change.